Add dumps from msm8996
Taken using:
strace -p 574 -f -xx -tt -s 2000 2>&1 /dev/null | grep AF_IB
(PID 574 is android.hardware.sensors@1.0-service)

proximity-dump is taken with the screen off, with a finger being moved in
front of the proximity sensor. accel-dump is taken with the screen on, where it
seems that the accelerometer, and possibly other sensors are turned on. Phone was
rotated around all 3 axes while taking accel-dump.